# swagger/docs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/auth/v1/service.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/training/v1/trainings": {
      "post": {
        "operationId": "TrainingService_CreateTraining",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateTrainingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateTrainingRequest"
            }
          }
        ],
        "tags": [
          "TrainingService"
        ]
      }
    },
    "/api/training/v1/trainings/{id}": {
      "get": {
        "operationId": "TrainingService_GetTraining",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetTrainingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "TrainingService"
        ]
      },
      "delete": {
        "operationId": "TrainingService_DeleteTraining",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteTrainingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "TrainingService"
        ]
      },
      "patch": {
        "operationId": "TrainingService_UpdateTraining",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateTrainingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TrainingServiceUpdateTrainingBody"
            }
          }
        ],
        "tags": [
          "TrainingService"
        ]
      }
    },
...
  },
  "definitions": {
    "TrainingServiceUpdateTrainingBody": {
      "type": "object",
      "properties": {
        "trainer_id": {
          "type": "string"
        },
        "user_id": {
          "type": "string"
        },
        "started_at": {
          "type": "string",
          "format": "date-time"
        },
        "ended_at": {
          "type": "string",
          "format": "date-time"
        },
        "additional_info": {
          "type": "string"
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateTrainingRequest": {
      "type": "object",
      "properties": {
        "trainer_id": {
          "type": "string"
        },
        "user_id": {
          "type": "string"
        },
        "started_at": {
          "type": "string",
          "format": "date-time",
          "title": "указываем полный путь до структуры"
        },
        "ended_at": {
          "type": "string",
          "format": "date-time"
        },
        "additional_info": {
          "type": "string"
        }
      }
    },
    "v1CreateTrainingResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "v1DeleteTrainingResponse": {
      "type": "object"
    },
...
    "v1GetTrainingResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "trainer_id": {
          "type": "string"
        },
        "user_id": {
          "type": "string"
        },
        "started_at": {
          "type": "string",
          "format": "date-time"
        },
        "ended_at": {
          "type": "string",
          "format": "date-time"
        },
        "additional_info": {
          "type": "string"
        }
      }
    },
...
    "v1UpdateTrainingResponse": {
      "type": "object"
    },
...
  }
}